D_F_A/robot-framework-tests/_archive_robot-tests/old_revisit/suites/requirements.txt
asn1crypto==1.4.0
backports.functools-lru-cache==1.6.4
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
#cffi==1.12.3
chardet==4.0.0
colorlog==6.4.1
cryptography==3.4.8
docutils >= 0.14enum34==1.1.6
et-xmlfile==1.1.0
future==0.18.2
idna==3.2
invoke >= 0.20
ipaddress==1.0.23
jdcal==1.4.1
jsonpatch==1.32
jsonpointer==2.1
junit-xml==1.9
natsort==7.1.1
ndg-httpsclient==0.5.1
numpy==1.21.2
o365==2.0.15
#o365==0.9.16
openpyxl==3.0.7
pandas==1.3.2
paramiko==2.7.2
Pillow==8.3.2
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.4.0
pyodbc==4.0.32
pyOpenSSL==20.0.1
PyYAML==5.4.1
rellu >= 0.6
requests==2.26.0
robotframework==4.1
robotframework-angularjs==0.0.10
robotframework-databaselibrary==1.2.4
#robotframework-excellibrary==0.0.2
robotframework-faker==5.0.0
robotframework-ftplibrary==1.9
robotframework-httplibrary==0.4.2
robotframework-pabot==2.0.1
robotframework-requests==0.9.1
robotframework-selenium2library==3.0.0
#robotframework-seleniumlibrary==3.3.1
robotframework-seleniumlibrary==5.1.3
robotframework-sshlibrary==3.7.0
robotframework-testrail==2.0.1
robotframework-xvfb==1.2.2
scp==0.14.0
selenium==3.141.0
setuptools
six==1.16.0
soupsieve==2.2.1
urllib3
waitress==2.0.0
WebOb==1.8.7
webdrivermanager
WebTest==3.0.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
xvfbwrapper==0.2.9
